Move recentFile-related code into new recent file
Issue #315
